finish webfn

+++ PythonCrashCourse/learning_log/learning_logs/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from django.urls import reverse
 from django.contrib.auth.decorators import login_required
 
@@ -14,18 +14,26 @@ def index(request):
 
 @login_required
 def topics(request):
-    topics = Topic.objects.order_by('date_added')
+    # print(str(request.user) == 'admin')
+    if str(request.user) == 'admin':
+        topics = Topic.objects.order_by('date_added')
+    else:
+        topics = Topic.objects.filter(owner=request.user).order_by('date_added')
     context = {'topics':topics}
     return render(request, 'learning_logs/topics.html', context)
 
 @login_required
 def topic(request, topic_id, entry_id=0):
+    topic = Topic.objects.get(id=topic_id)
+    if topic.owner != request.user and str(request.user) != 'admin':
+        raise Http404
+
     if 'delete' in request.path_info and Entry.objects.get(id=entry_id):
         entry = Entry.objects.get(id=entry_id)
         topic = entry.topic 
         entry.delete()
         return HttpResponseRedirect(reverse('learning_logs:topic',args=[topic_id]))
-    topic = Topic.objects.get(id=topic_id)
+    
     entries = topic.entry_set.order_by('-date_added')
     context = {'topic':topic, 'entries':entries}
     return render(request, 'learning_logs/topic.html', context)
@@ -38,7 +46,10 @@ def new_topic(request):
 
         form = TopicForm(request.POST)
         if form.is_valid():
-            form.save()
+            new_topic = form.save(commit=False)
+            new_topic.owner = request.user
+            new_topic.save()
+            
             return HttpResponseRedirect(reverse('learning_logs:topics'))
 
     context = {'form': form}
@@ -68,6 +79,8 @@ def edit_entry(request, entry_id):
     # if  "delete" in request.path_info:
     #     entry.delete()
     #     return HttpResponseRedirect(reverse('learning_logs:topic',args=[topic.id]))
+    if topic.owner != request.user and str(request.user) != 'admin':
+        raise Http404
     if request.method != 'POST':
         form = EntryForm(instance=entry)
     else:
